requirements.txt
streamlit
llama-cpp-python
huggingface_hub
datasets